+++ src/lib/ElementDescriptorContainer.js
@@ -14,6 +14,7 @@ import CubeCameraDescriptor from './descriptors/Object/Camera/CubeCameraDescript
 
 import MeshDescriptor from './descriptors/Object/MeshDescriptor';
 import LineDescriptor from './descriptors/Object/LineDescriptor';
+import LineSegmentsDescriptor from './descriptors/Object/LineSegmentsDescriptor';
 import PointsDescriptor from './descriptors/Object/PointsDescriptor';
 import SpriteDescriptor from './descriptors/Object/SpriteDescriptor';
 
@@ -105,6 +106,7 @@ class ElementDescriptorContainer {
 
       mesh: new MeshDescriptor(react3RendererInstance),
       line: new LineDescriptor(react3RendererInstance),
+      lineSegments: new LineSegmentsDescriptor(react3RendererInstance),
       points: new PointsDescriptor(react3RendererInstance),
       sprite: new SpriteDescriptor(react3RendererInstance),